Integer ids replaced by Long ids.

+++ neo4jtools-objectgraphmapper/src/test/java/de/christofreichardt/neo4jtools/ogm/model/EncryptedDocument.java
@@ -9,7 +9,7 @@ import de.christofreichardt.neo4jtools.apt.NodeEntity;
 @NodeEntity(label = "DOCUMENTS")
 public class EncryptedDocument extends Document {
 
-  public EncryptedDocument(Integer id) {
+  public EncryptedDocument(Long id) {
     super(id);
   }
 }
