+++ neo4jtools-objectgraphmapper/src/test/java/de/christofreichardt/neo4jtools/ogm/model/PlaintextDocument.java
@@ -9,7 +9,7 @@ import de.christofreichardt.neo4jtools.apt.NodeEntity;
 @NodeEntity(label = "DOCUMENTS")
 public class PlaintextDocument extends Document {
 
-  public PlaintextDocument(Integer id) {
+  public PlaintextDocument(Long id) {
     super(id);
   }
 }
